flaskapp/api/requirements.txt
flask==2.0.1
flask-cors
Werkzeug==2.2.2
numpy
opencv-python
Pillow
tensorflow
